[ai][js] remove dead Overpass map module from web editor
The map editor is tile-snapshot + upload only, so the Overpass vector
generation path (generateMapSvg, buildOverpassQueries, categoriseElements,
makeCoordConverters, …) and the fetchOverpassQuery proxy client are no
longer reachable. Delete map-svg.ts and map-svg.test.ts, drop
fetchOverpassQuery/OverpassResult from api.ts, and move the one survivor —
svgToBase64 — into lib/svg.ts with its own focused test.
